V1.1.2
Fix a couple of bugs, added the altitude, added the artifacts, etc

+++ src/buildings/FuelBuilding.h
@@ -48,7 +48,7 @@ class FuelBuilding : public Building
             void exitButton() { m_wantToExit = true; }
 
         // Some variables
-            const int m_fuelPrice;
+            const float m_fuelPrice;
             bool m_wantToExit;
             Character *m_ptrCharacter;
